# backend/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Teranga API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/providers": {
      "post": {
        "summary": "Submit a provider application (role 'provider' only)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProviderRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Provider profile created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "403": {
            "description": "Caller is not a provider-role account",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List providers within the caller's scope (admin, category_manager)",
        "responses": {
          "200": {
            "description": "Scoped provider list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderListResponse"
                }
              }
            }
          },
          "403": {
            "description": "Caller has no manageable scope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/missions": {
      "post": {
        "summary": "Create a guided mission for the authenticated client (role: client)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateMissionRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mission created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MissionResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid payload or unresolvable address",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/saved-locations": {
      "get": {
        "summary": "List the authenticated client's saved locations (role: client)",
        "responses": {
          "200": {
            "description": "Saved locations",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SavedLocationListResponse" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Save a favorite location (role: client)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateSavedLocationRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Saved location created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SavedLocationResponse" }
              }
            }
          }
        }
      }
    },
    "/mission-pricing-rules": {
      "get": {
        "summary": "List pricing rules within the caller's scope (admin, global or country/region-scoped)",
        "responses": {
          "200": {
            "description": "Scoped pricing rules",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MissionPricingRuleListResponse" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a pricing rule (admin, global or country/region-scoped)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateMissionPricingRuleRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pricing rule created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MissionPricingRuleResponse" }
              }
            }
          },
          "403": {
            "description": "Caller has no manageable scope for this country/region",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
...
      "Provider": {
        "type": "object",
        "required": ["id", "userId", "type", "displayFirstName", "status"],
        "properties": {
          "id": { "type": "number" },
          "userId": { "type": "number" },
          "type": { "type": "string", "enum": ["independent", "company"] },
          "legalName": { "type": ["string", "null"] },
          "displayFirstName": { "type": "string" },
          "rccmNumber": { "type": ["string", "null"] },
          "phoneNumber": { "type": "string" },
          "email": { "type": ["string", "null"] },
          "countryCode": { "type": "string" },
          "status": {
            "type": "string",
            "enum": ["pending", "probation", "active", "suspended", "revoked"]
          },
          "averageRating": { "type": ["number", "null"] },
          "completedMissionsCount": { "type": "number" },
          "hasLiabilityInsurance": { "type": "boolean" },
          "badgeCertified": { "type": "boolean" }
        },
        "additionalProperties": true
      },
      "ProviderResponse": {
        "type": "object",
        "required": ["provider"],
        "properties": {
          "provider": { "$ref": "#/components/schemas/Provider" }
        },
        "additionalProperties": true
      },
      "ProviderListResponse": {
        "type": "object",
        "required": ["providers", "pagination"],
        "properties": {
          "providers": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Provider" }
          },
          "pagination": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "additionalProperties": true
      },
      "CreateProviderRequest": {
        "type": "object",
        "required": ["type", "displayFirstName", "phoneNumber", "countryCode", "tradeCategoryIds"],
        "properties": {
          "type": { "type": "string", "enum": ["independent", "company"] },
          "legalName": { "type": "string" },
          "displayFirstName": { "type": "string" },
          "rccmNumber": { "type": "string" },
          "phoneNumber": { "type": "string" },
          "email": { "type": "string" },
          "countryCode": { "type": "string", "minLength": 2, "maxLength": 2 },
          "hasLiabilityInsurance": { "type": "boolean" },
          "insuranceExpiresAt": { "type": "string", "format": "date" },
          "tradeCategoryIds": {
            "type": "array",
            "items": { "type": "number" },
            "minItems": 1
          }
        },
        "additionalProperties": false
      },
...
      "MissionEstimate": {
        "type": "object",
        "required": ["pricingMode", "currency", "estimatedDelayMinutes"],
        "properties": {
          "pricingMode": { "type": "string", "enum": ["fixed_estimate", "quote_only"] },
          "currency": { "type": "string" },
          "basePrice": { "type": ["number", "null"] },
          "minPrice": { "type": ["number", "null"] },
          "pricePerKm": { "type": "number" },
          "estimatedDelayMinutes": { "type": "number" },
          "note": { "type": "string" }
        },
        "additionalProperties": true
      },
...
      "CreateMissionRequest": {
        "type": "object",
        "required": ["executionType", "title"],
        "properties": {
          "executionType": { "type": "string", "enum": ["agent", "provider"] },
          "tradeCategoryId": { "type": "number" },
          "serviceType": {
            "type": "string",
            "enum": ["errand", "administrative", "payment", "money_transfer", "other"]
          },
          "title": { "type": "string", "minLength": 3, "maxLength": 150 },
          "description": { "type": "string", "maxLength": 2000 },
          "savedLocationId": { "type": "number" },
          "address": { "type": "string", "maxLength": 255 },
          "latitude": { "type": "number" },
          "longitude": { "type": "number" }
        },
        "additionalProperties": false
      },
      "MissionResponse": {
        "type": "object",
        "required": ["message", "mission", "estimate"],
        "properties": {
          "message": { "type": "string" },
          "mission": { "type": "object", "additionalProperties": true },
          "estimate": { "$ref": "#/components/schemas/MissionEstimate" }
        },
        "additionalProperties": true
      },
...
      "SavedLocation": {
        "type": "object",
        "required": ["id", "userId", "address", "latitude", "longitude"],
        "properties": {
          "id": { "type": "number" },
          "userId": { "type": "number" },
          "label": { "type": ["string", "null"] },
          "address": { "type": "string" },
          "latitude": { "type": "number" },
          "longitude": { "type": "number" }
        },
        "additionalProperties": true
      },
      "SavedLocationListResponse": {
        "type": "object",
        "required": ["savedLocations"],
        "properties": {
          "savedLocations": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SavedLocation" }
          }
        },
        "additionalProperties": true
      },
      "SavedLocationResponse": {
        "type": "object",
        "required": ["savedLocation"],
        "properties": {
          "savedLocation": { "$ref": "#/components/schemas/SavedLocation" }
        },
        "additionalProperties": true
      },
      "CreateSavedLocationRequest": {
        "type": "object",
        "required": ["address", "latitude", "longitude"],
        "properties": {
          "label": { "type": "string", "maxLength": 80 },
          "address": { "type": "string", "maxLength": 255 },
          "latitude": { "type": "number", "minimum": -90, "maximum": 90 },
          "longitude": { "type": "number", "minimum": -180, "maximum": 180 }
        },
        "additionalProperties": false
      },
      "MissionPricingRule": {
        "type": "object",
        "required": ["id", "countryId", "pricingMode", "estimatedDelayMinutes"],
        "properties": {
          "id": { "type": "number" },
          "countryId": { "type": "number" },
          "regionId": { "type": ["number", "null"] },
          "tradeCategoryId": { "type": ["number", "null"] },
          "serviceType": { "type": ["string", "null"] },
          "pricingMode": { "type": "string", "enum": ["fixed_estimate", "quote_only"] },
          "basePrice": { "type": ["number", "null"] },
          "minPrice": { "type": ["number", "null"] },
          "pricePerKm": { "type": "number" },
          "estimatedDelayMinutes": { "type": "number" },
          "isActive": { "type": "boolean" }
        },
        "additionalProperties": true
      },
      "MissionPricingRuleListResponse": {
        "type": "object",
        "required": ["pricingRules"],
        "properties": {
          "pricingRules": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/MissionPricingRule" }
          }
        },
        "additionalProperties": true
      },
      "MissionPricingRuleResponse": {
        "type": "object",
        "required": ["pricingRule"],
        "properties": {
          "pricingRule": { "$ref": "#/components/schemas/MissionPricingRule" }
        },
        "additionalProperties": true
      },
      "CreateMissionPricingRuleRequest": {
        "type": "object",
        "required": ["countryId", "pricingMode", "estimatedDelayMinutes"],
        "properties": {
          "countryId": { "type": "number" },
          "regionId": { "type": "number" },
          "tradeCategoryId": { "type": "number" },
          "serviceType": {
            "type": "string",
            "enum": ["errand", "administrative", "payment", "money_transfer", "other"]
          },
          "pricingMode": { "type": "string", "enum": ["fixed_estimate", "quote_only"] },
          "basePrice": { "type": "number", "minimum": 0 },
          "minPrice": { "type": "number", "minimum": 0 },
          "pricePerKm": { "type": "number", "minimum": 0 },
          "estimatedDelayMinutes": { "type": "number", "minimum": 0 }
        },
        "additionalProperties": false
      },
...
    }
  }
}